backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0
requests==2.31.0
spotipy==2.23.0
joblib==1.3.2
asyncpg==0.29.0
celery==5.3.4
flower==2.0.1
prometheus-client==0.19.0
structlog==23.2.0 